moved processing module into stimuli/preprocessing
