+++ examples/stimuli/plot_retarget.py
