--- pulse2percept/processing/__init__.py
@@ -1,19 +1,0 @@
-"""Processing techniques that can be used to process image or video before it becomes stimuli
-.. autosummary::
-    :toctree: _api
-
-    scene_retargeting
-..
-
-"""
-
-from .scene_retargeting import single_image_retargeting, image_retargeting, video_retargeting, stim_retargeting, _spatial_temporal_saliency
-
-
-__all__ = [
-    'single_image_retargeting',
-    'image_retargeting', 
-    'video_retargeting',
-    'stim_retargeting', 
-    '_spatial_temporal_saliency'
-]
