+++ pulse2percept/stimuli/preprocessing.py
