+++ pulse2percept/stimuli/setup.py
@@ -16,6 +16,10 @@ def configuration(parent_package='', top_path=None):
                              sources=['_base.pyx'],
                              include_dirs=[numpy.get_include()],
                              libraries=libraries)
+        config.add_extension('_preprocessing',
+                             sources=['_preprocessing.pyx'],
+                             include_dirs=[numpy.get_include()],
+                             libraries=libraries)
     config.add_subpackage('tests')
     config.add_data_dir('data')
